+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Fixed
+- Add missing CreateAssetMenu attribute in PersistenceSettings
 - Add missing namespace for NewtonsoftSerializer
 
 ## [2.1.1] - 2023-03-15