+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.2.0] - 2023-03-28
 ### Added
 - Fast load raw file
 - Check if generic type is serializable before load/save
@@ -64,7 +66,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - gitignore
 - Initial commit
 
-[Unreleased]: https://github.com/HyagoOliveira/Persistence/compare/2.1.1...main
+[Unreleased]: https://github.com/HyagoOliveira/Persistence/compare/2.2.0...main
+[2.2.0]: https://github.com/HyagoOliveira/Persistence/tree/2.2.0/
 [2.1.1]: https://github.com/HyagoOliveira/Persistence/tree/2.1.1/
 [2.1.0]: https://github.com/HyagoOliveira/Persistence/tree/2.1.0/
 [2.0.0]: https://github.com/HyagoOliveira/Persistence/tree/2.0.0/